requirements.txt
altair==4.2.2
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blinker==1.5
bytetracker==0.3.2
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.0
contourpy==1.0.7
crispy-bootstrap5==0.7
cron-descriptor==1.2.35
cycler==0.11.0
decorator==5.1.1
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cleanup==7.0.0
django-crispy-forms==2.0
django-debug-toolbar==3.8.1
django-timezone-field==5.0
entrypoints==0.4
filelock==3.10.0
filterpy==1.4.5
fonttools==4.39.2
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.27.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
kombu==5.2.4
lap==0.4.0
lazy-loader==0.2
lit==15.0.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.0
pafy==0.5.5
pandas==1.5.3
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pycodestyle==2.10.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
rich==13.3.2
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.12.2
semver==2.13.0
sentry-sdk==1.17.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
streamlit-nightly==1.20.1.dev20230316
supervision==0.3.2
sympy==1.11.1
thop==0.1.1.post2209072238
tifffile==2023.3.21
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ultralytics==8.0.54
urllib3==1.26.15
validators==0.20.0
vine==5.0.0
watchdog==2.3.1
wcwidth==0.2.6
youtube-dl==2020.12.2
zipp==3.15.0
